attempt ref path
