+++ packages/server/server.ts
@@ -13,7 +13,6 @@ import jiraImagesHandler from './jiraImagesHandler'
 import listenHandler from './listenHandler'
 import selfHostedHandler from './selfHostedHandler'
 import {createStaticFileHandler} from './staticFileHandler'
-import './types/webpackEnv.d'
 import {Logger} from './utils/Logger'
 import SAMLHandler from './utils/SAMLHandler'
 import getReqAuth from './utils/getReqAuth'
